003_orchestration/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas 
sqlalchemy
#psycopg2-binary
pyarrow
psycopg2-binary==2.9.10
boto3


airflow-provider-mlflow
apache-airflow-providers-amazon
mlflow-skinny
astro-sdk-python